+++ packages/wallet/src/initialization/instances/transaction-controller/transaction-controller.test.ts
@@ -1,4 +1,3 @@
-import { Env } from '@metamask/claims-controller';
 import { Messenger } from '@metamask/messenger';
 import { InMemoryStorageAdapter } from '@metamask/storage-service';
 import { TransactionController } from '@metamask/transaction-controller';
@@ -126,10 +125,6 @@ function getInstanceOptions(): WalletOptions['instanceOptions'] {
       storage: new InMemoryStorageAdapter(),
     },
     remoteFeatureFlagController: REMOTE_FEATURE_FLAG_OPTIONS,
-    claimsService: {
-      env: Env.DEV,
-      fetchFunction: globalThis.fetch,
-    },
   };
 }